# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"ZamTrust API","description":"Secure Digital Document Signing and Verification Platform. Provides document upload, cryptographic hashing (SHA-256), RSA-2048 digital signing, verification, QR codes, and audit trails.","contact":{"name":"ZamTrust Project","email":"support@zamtrust.example"},"license":{"name":"Academic Use — MSc Computer Science"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local development"}],"security":[{"bearerAuth":[]}],"paths":{"/api/verifications/{verificationId}/upload":{"post":{"tags":["verification-controller"],"operationId":"verifyUpload","parameters":[{"name":"verificationId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResult"}}}}}}},"/api/documents":{"get":{"tags":["document-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentResponse"}}}}}}},"post":{"tags":["document-controller"],"operationId":"upload","parameters":[{"name":"title","in":"query","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}}}}},"/api/documents/{id}/sign":{"post":{"tags":["document-controller"],"operationId":"sign","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignatureResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/verifications/{verificationId}":{"get":{"tags":["verification-controller"],"operationId":"verify","parameters":[{"name":"verificationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VerificationResult"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/public/qr/{verificationId}":{"get":{"tags":["qr-controller"],"operationId":"qr","parameters":[{"name":"verificationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"image/png":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/documents/{id}":{"get":{"tags":["document-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}}}}},"/api/audit/user/{userId}":{"get":{"tags":["audit-controller"],"operationId":"byUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditLog"}}}}}}}},"/api/audit/resource/{resource}":{"get":{"tags":["audit-controller"],"operationId":"byResource","parameters":[{"name":"resource","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditLog"}}}}}}}},"/api/audit/all":{"get":{"tags":["audit-controller"],"operationId":"all","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditLog"}}}}}}}}},"components":{"schemas":{"VerificationResult":{"type":"object","properties":{"verificationId":{"type":"string"},"fileName":{"type":"string"},"integrityValid":{"type":"boolean"},"signatureValid":{"type":"boolean"},"signer":{"type":"string"},"signedAt":{"type":"string","format":"date-time"},"algorithm":{"type":"string"},"message":{"type":"string"}}},"DocumentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"verificationId":{"type":"string"},"fileName":{"type":"string"},"title":{"type":"string"},"contentType":{"type":"string"},"sizeBytes":{"type":"integer","format":"int64"},"originalHash":{"type":"string"},"status":{"type":"string"},"ownerUsername":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"signedAt":{"type":"string","format":"date-time"}}},"SignatureResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"documentId":{"type":"integer","format":"int64"},"signerUsername":{"type":"string"},"algorithm":{"type":"string"},"documentHash":{"type":"string"},"signedAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","password","username"],"type":"object","properties":{"username":{"maxLength":64,"minLength":3,"type":"string"},"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"},"fullName":{"type":"string"},"organization":{"type":"string"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"passwordHash":{"type":"string"},"fullName":{"type":"string"},"organization":{"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["SYSTEM_ADMIN","ORG_ADMIN","SIGNER","DOCUMENT_USER","VERIFIER"]}},"enabled":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"AuditLog":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"action":{"type":"string"},"resource":{"type":"string"},"details":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"ipAddress":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Paste your JWT token here. Obtain it via POST /api/auth/login.","scheme":"bearer","bearerFormat":"JWT"}}}}